tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-flow-valid.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Grid Lanes: parsing grid-lanes-flow with valid values</title>
  <link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
  <script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
  <script>
    test_valid_value("grid-lanes-flow", "column normal");
    test_valid_value("grid-lanes-flow", "column reverse");
    test_shorthand_value('grid-lanes-flow', 'column normal', {
      'grid-lanes-direction': 'column',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes-flow', 'column reverse', {
      'grid-lanes-direction': 'column',
      'grid-lanes-fill': 'reverse'
    });

    test_valid_value("grid-lanes-flow", "row normal");
    test_valid_value("grid-lanes-flow", "row reverse");
    test_shorthand_value('grid-lanes-flow', 'row normal', {
      'grid-lanes-direction': 'row',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes-flow', 'row reverse', {
      'grid-lanes-direction': 'row',
      'grid-lanes-fill': 'reverse'
    });

    test_valid_value("grid-lanes-flow", "column-reverse normal");
    test_valid_value("grid-lanes-flow", "column-reverse reverse");
    test_shorthand_value('grid-lanes-flow', 'column-reverse normal', {
      'grid-lanes-direction': 'column-reverse',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes-flow', 'column-reverse reverse', {
      'grid-lanes-direction': 'column-reverse',
      'grid-lanes-fill': 'reverse'
    });

    test_valid_value("grid-lanes-flow", "row-reverse normal");
    test_valid_value("grid-lanes-flow", "row-reverse reverse");
    test_shorthand_value('grid-lanes-flow', 'row-reverse normal', {
      'grid-lanes-direction': 'row-reverse',
      'grid-lanes-fill': 'normal'
    });
    test_shorthand_value('grid-lanes-flow', 'row-reverse reverse', {
      'grid-lanes-direction': 'row-reverse',
      'grid-lanes-fill': 'reverse'
    });
  </script>
</body>
</html>
